+++ khorosjx/content/ideas.py
@@ -6,12 +6,12 @@
 :Example:           ``content_id = ideas.get_content_id(url)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     04 Mar 2020
+:Modified Date:     16 Mar 2020
 """
 
 from .. import core
 from . import base
-from ..utils import core_utils
+from ..utils import core_utils, df_utils
 
 
 # Define function to verify the connection in the core module
@@ -19,13 +19,13 @@ def verify_core_connection():
     """This function verifies that the core connection information (Base URL and API credentials) has been defined.
 
     :returns: None
-    :raises: NameError, KhorosJXError, NoCredentialsError
+    :raises: :py:exc:`NameError`, KhorosJXError, NoCredentialsError
     """
     def __get_info():
         """This function initializes and defines the global variables for the connection information.
 
         :returns: None
-        :raises: NameError, KhorosJXError, NoCredentialsError
+        :raises: :py:exc:`NameError`, KhorosJXError, NoCredentialsError
         """
         # Initialize global variables
         global base_url
@@ -50,13 +50,13 @@ def get_content_id(url):
     :param url: The URL of the idea
     :type url: str
     :returns: The Content ID for the idea
-    :raises: ValueError
+    :raises: :py:exc:`ValueError`
     """
     content_id = base.get_content_id(url, 'idea')
     return content_id
 
 
-def get_ideas_for_space(browse_id):
+def get_ideas_for_space(browse_id, return_type='list', ignore_exceptions=False):
     # Verify that the core connection has been established
     verify_core_connection()
 
@@ -68,13 +68,17 @@ def get_ideas_for_space(browse_id):
     endpoint = f"places/{browse_id}/contents"
     query_string = "filter=type(idea)"
     ideas = base.get_paginated_content(endpoint, query_string, start_index, 'idea', all_fields=True,
-                                       ignore_exceptions=True)
+                                       ignore_exceptions=ignore_exceptions)
     all_ideas = core_utils.add_to_master_list(ideas, all_ideas)
 
     # Continue querying for groups until none are returned
     while len(ideas) > 0:
         start_index += 100
         ideas = base.get_paginated_content(endpoint, query_string, start_index, 'idea', all_fields=True,
-                                        ignore_exceptions=True)
+                                           ignore_exceptions=ignore_exceptions)
         all_ideas = core_utils.add_to_master_list(ideas, all_ideas)
-    
+
+    # Return the data as a master list of group dictionaries or a pandas dataframe
+    if return_type == "dataframe":
+        all_ideas = df_utils.convert_dict_list_to_dataframe(all_ideas)
+    return all_ideas